requirements.txt
Crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
PyYAML==6.0.1
groq>=0.13.0
